[FEAT] 시스템 관리자 페이지 카페 관리 관련 조회 로직 코드 작성.

+++ src/main/java/com/grinder/common/model/PageVO.java
@@ -15,7 +15,7 @@ public class PageVO implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private int page;
+    private long page;
 
-    private int size;
+    private long size;
 }
